feat: qna 조회 및 질문 작성 기능

+++ backend/src/main/java/com/chestnut/backend/qna/repository/QnACategoryRepository.java
@@ -3,6 +3,6 @@ package com.chestnut.backend.qna.repository;
 import com.chestnut.backend.qna.entity.QnACategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QnACategoryRepository extends JpaRepository<QnACategory, Long> {
+public interface QnACategoryRepository extends JpaRepository<QnACategory, Byte> {
 
 }
